Move features config to the backend
