+++ apps/design/backend/src/index.ts
@@ -17,6 +17,7 @@ import {
 
 export type { ElectionRecord } from './store';
 export type { BallotOrderInfo, BallotStyle, User } from './types';
+export type { ElectionFeaturesConfig, UserFeaturesConfig } from './features';
 export type { Api, ElectionInfo } from './app';
 
 export type { BallotMode } from '@votingworks/hmpb';
